+++ proses_genre.php
@@ -41,9 +41,23 @@ if (isset($_POST['simpan'])) {
 		</script>";
     }
 } else {
-    include 'connection.php';
 
     $id = $_GET['id'];
+
+    $sql = "SELECT * FROM `film` 
+    WHERE `film`.`id_genre` = '$id'";
+    $cek_genre = mysqli_query($conn, $sql);
+
+    $count_genre = mysqli_fetch_assoc($cek_genre);
+    if ($count_genre > 0) {
+        echo "
+        <script>
+                alert('Sudah Ada Film Yang Menggunakan Genre ini,Data Genre Tidak Dapat Di Hapus');
+                window.location.href='master_genre.php?id=$id';
+        </script>";
+        die;
+    }
+
     $sql = "DELETE FROM `genre` WHERE id_genre=$id";
     $hapus = mysqli_query($conn, $sql);
